infrastructure/src/main/java/com/mercadolibre/planning/model/me/controller/deviation/DeviationController.java
package com.mercadolibre.planning.model.me.controller.deviation;

import com.mercadolibre.planning.model.me.controller.deviation.request.DeviationRequest;
import com.mercadolibre.planning.model.me.controller.editor.WorkflowEditor;
import com.mercadolibre.planning.model.me.gateways.planningmodel.dtos.DeviationResponse;
import com.mercadolibre.planning.model.me.gateways.planningmodel.dtos.Workflow;
import com.mercadolibre.planning.model.me.usecases.authorization.AuthorizeUser;
import com.mercadolibre.planning.model.me.usecases.authorization.dtos.AuthorizeUserDto;
import com.mercadolibre.planning.model.me.usecases.deviation.SaveDeviation;
import com.newrelic.api.agent.Trace;
import lombok.AllArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.PropertyEditorRegistry;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;

import org.springframework.web.bind.annotation.InitBinder;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;

import static com.mercadolibre.planning.model.me.gateways.authorization.dtos.UserPermission.OUTBOUND_SIMULATION;
import static java.util.Collections.singletonList;

@Slf4j
@AllArgsConstructor
@RestController
@RequestMapping("/planning/model/middleend/workflows/{workflow}/deviations")
public class DeviationController {

    private SaveDeviation saveDeviation;
    private final AuthorizeUser authorizeUser;

    @Trace
    @PostMapping("/save")
    public ResponseEntity<DeviationResponse> save(
            @PathVariable final Workflow workflow,
            @RequestBody @Valid final DeviationRequest deviationRequest) {

        authorizeUser.execute(new AuthorizeUserDto(
                deviationRequest.getUserId(), singletonList(OUTBOUND_SIMULATION)));

        return ResponseEntity.status(HttpStatus.OK)
                .body(saveDeviation.execute(
                        deviationRequest.toDeviationInput(workflow)));
    }

    @InitBinder
    public void initBinder(final PropertyEditorRegistry dataBinder) {
        dataBinder.registerCustomEditor(Workflow.class, new WorkflowEditor());
    }
}
